+++ test/validation/dry_validation_test.rb
@@ -533,7 +533,7 @@ class ValidationGroupsTest < MiniTest::Spec
       end
 
       # run this is :email group is true.
-      validation name: :after_email, if: lambda { |results| results[:email]==true } do # extends the above.
+      validation name: :after_email, if: lambda { |results| results[:email].success? } do # extends the above.
         required(:username).filled
       end