+++ CHANGES.md
@@ -19,6 +19,10 @@ You can upgrade from 2.2.0 without worries.
 * [private] `Group#call` API now is: `call(form, errors)`.
 * Removed `Form#valid?`.
 
+* In `:if` for validation groups, you now get a hash of result objects, not just true/false.
+
+### TODO: deprecate that.
+
 
 ## 2.2.4